feat(api): moved event listener to specific service
